pyproject.toml
[tool.poetry]
name = "project"
version = "0.1.0"
description = "this is the boilerplate code for django and react project"
authors = ["Gaurav Dhiman <gaurav.d@idsil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
djangorestframework-simplejwt = "^5.3.1"
pre-commit = "^4.0.0"
gunicorn = "^23.0.0"
python-dotenv = "^1.0.1"
dj-database-url = "^2.2.0"
redis = "^5.1.1"
celery = "^5.4.0"
ruff = "^0.6.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
